Split pure audio timing/SSML helpers out of audio_edit_tools
generate_ssml_line, text_to_keypoints, timings_to_text and
timing_text_without_filename move verbatim to audio_timing.ts;
audio_edit_tools re-exports them for existing consumers. The story parser now
imports from the pure module, so it no longer pulls CodeMirror and can be
bundled server-side (Convex node actions).
